feat(server): remove duplicate server, pass flag for local vs observer
Before:

Local    - `python camera/server/local.py`
Observer - `python camera/server/observer.py`

After:

Local    - `python camera/server/server.py local`
Observer - `python camera/server/server.py observer`
            (you can omit `observer`)
